backend/requirements.txt
# Flask and related
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-PyMongo==2.3.0
pymongo==4.6.0
dnspython==2.6.1
bcrypt==4.1.2
python-dotenv==1.0.0

# Core ML / numerical libraries
numpy==1.26.4           # compatible with concrete-ml 1.9.0
pandas==2.0.3        # works with numpy 1.26.x
scikit-learn==1.5.0     # pinned by concrete-ml
scipy==1.11.4
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Torch stack (compatible with concrete-ml 1.9.0)
torch==2.3.1
torchvision==0.18.1    # matches torch 2.3.1

# Reinforcement learning
gym==0.26.2             # compatible with Python 3.11+
stable-baselines3==2.2.1

# Concrete-ML
concrete-ml==1.9.0

# Cloud / crypto / web
cryptography==41.0.7
boto3==1.34.0
dash==2.14.2
dash-bootstrap-components==1.5.0
